+++ src/model/lstm.py
@@ -17,8 +17,8 @@ class LSTM(AnomalyDetector):
     ):
         super(LSTM, self).__init__(**hyperparams)
         self.lstm = torch.nn.LSTM(input_size, hidden_state, num_layers=n_layers, bidirectional=bidirectional,
-                                  batch_first=True, dropout=dropout).to(torch.float64).to(device)
-        self.ff = torch.nn.Linear(hidden_state, input_size).to(torch.float64).to(device)
+                                  batch_first=True, dropout=dropout).to(torch.float32).to(device)
+        self.ff = torch.nn.Linear(hidden_state, input_size).to(torch.float32).to(device)
 
     def forward(self, x):
         out, _ = self.lstm(x)
